requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
sentence-transformers==2.2.2
oracledb==2.2.1
fastapi==0.111.0
rich==13.7.1
Werkzeug==3.0.3
markdown==3.6
python-dotenv==1.0.1
oci==2.129.3
langchain_community==0.2.9
docx2txt==0.8
langchain-huggingface==0.0.3